requirements.txt
amqp==5.1.1
APScheduler==3.6.3
asgiref==3.4.1
async-generator==1.10
async-timeout==4.0.2
atomicwrites==1.4.0
attrs==21.4.0
autobahn==23.1.1
Automat==22.10.0
billiard==3.6.4.0
black==23.3.0
cachetools==4.2.2
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==2.0.12
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
constantly==15.1.0
coverage==7.2.7
cron-descriptor==1.2.35
cryptography==39.0.0
cytoolz==0.11.2
daphne==4.0.0
distlib==0.3.6
dj-database-url==2.1.0
Django==3.2.8
django-celery-beat==2.5.0
django-cors-headers==3.10.0
django-timezone-field==5.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.5.0
emoji==2.2.0
eth-hash==0.3.2
eth-keys==0.3.4
eth-typing==2.2.2
eth-utils==1.10.0
exceptiongroup==1.1.0
execnet==1.9.0
filelock==3.9.0
frozenlist==1.3.0
h11==0.14.0
hexbytes==0.2.2
hyperlink==21.0.0
idna==3.3
incremental==22.10.0
iniconfig==1.1.1
kombu==5.2.4
msgpack==1.0.4
multidict==6.0.2
mypy-extensions==0.4.3
netaddr==0.8.0
outcome==1.2.0
packaging==23.0
parsimonious==0.8.1
pathspec==0.9.0
Pillow==9.4.0
pipenv==2023.2.4
platformdirs==2.4.1
pluggy==1.0.0
prompt-toolkit==3.0.38
psycopg2==2.9.5
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.14.1
PyJWT==1.7.1
pyOpenSSL==23.0.0
pyparsing==3.0.6
PySocks==1.7.1
pytest==6.2.5
pytest-forked==1.4.0
pytest-xdist==2.5.0
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==13.14
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
redis==4.4.2
requests==2.28.1
rlp==2.0.1
selenium==4.8.0
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.2
stripe==5.4.0
toml==0.10.2
tomli==1.2.3
toolz==0.11.1
tornado==6.1
trio==0.22.0
trio-websocket==0.9.2
Twisted==22.10.0
twisted-iocpsupport==1.0.2
txaio==23.1.1
typing_extensions==4.0.1
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.9
varint==1.0.2
vine==5.0.0
virtualenv==20.17.1
virtualenv-clone==0.5.7
virtualenvwrapper-win==1.2.7
wcwidth==0.2.6
wsproto==1.2.0
zope.interface==5.5.2
